requirements.txt
# Image Colorization Project Requirements
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0  # Alternative to PyTorch if needed
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
pillow>=10.0.0

# Web Framework
fastapi>=0.103.0
uvicorn>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Data Processing
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
jupyter>=1.0.0
notebook>=7.0.0